\name{createLowerMShipMatrix}
\alias{createLowerMShipMatrix}
\title{Create Lower Approximation}
\usage{
createLowerMShipMatrix(upperMShipMatrix)
}
\arguments{
  \item{upperMShipMatrix}{An upper approximation matrix.}
}
\value{
Returns the corresponding lower approximation.
}
\description{
Creates a lower approximation out of an upper
approximation.
}
\author{
G. Peters.
}

