% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoughClustering.r
\name{datatypeInteger}
\alias{datatypeInteger}
\title{Rough k-Means Plotting}
\usage{
datatypeInteger(x)
}
\arguments{
\item{x}{As a replacement for is.integer(). is.integer() delivers FALSE when the variable is numeric (as superset for integer etc.)}
}
\value{
TRUE if x is integer otherwise FALSE.
}
\description{
Checks for integer.
}
\author{
G. Peters.
}
