\name{DemoDataC2D2a}
\alias{DemoDataC2D2a}
\docType{data}
\title{A Simple Data Set}
\description{
This is a simple two-dimensional data set with 200 objects in 2 clusters. It is provided as sample data set to easily test the algorithms.
}
\usage{data(DemoDataC2D2a)}
\format{
  A data frame with 200 observations on 2 variables.
}
\examples{
data(DemoDataC2D2a); plot(DemoDataC2D2a)
}
\keyword{datasets}
