\name{initMeansC2D2a}
\alias{initMeansC2D2a}
\docType{data}
\title{Initial Means (2 Clusters)}
\description{
This file contains initial means for an example for the HardKMEANSDemo function.
}
\usage{data(initMeansC2D2a)}
\format{
  A data frame with 200 observations on 2 variables.
}
\examples{
data(initMeansC2D2a); plot(initMeansC2D2a)
}
\keyword{datasets}
