\name{initMeansC3D2a}
\alias{initMeansC3D2a}
\docType{data}
\title{Initial Means (3 Clusters)}
\description{
This file contains initial means for an example for the HardKMEANSDemo function.
}
\usage{data(initMeansC3D2a)}
\format{
  A matrix [nClusters x nFeatures] = [3 x 2].
}
\examples{
data(initMeansC3D2a); plot(initMeansC3D2a)
}
\keyword{datasets}
