\name{initMeansC4D2a}
\alias{initMeansC4D2a}
\docType{data}
\title{Initial Means (4 Clusters)}
\description{
This file contains initial means for an example for the HardKMEANSDemo function.
}
\usage{data(initMeansC4D2a)}
\format{
  A matrix [nClusters x nFeatures] = [4 x 2].
}
\examples{
data(initMeansC4D2a); plot(initMeansC4D2a)
}
\keyword{datasets}
