\name{initMeansC5D2a}
\alias{initMeansC5D2a}
\docType{data}
\title{Initial Means (5 Clusters)}
\description{
This file contains initial means for an example for the HardKMEANSDemo function.
}
\usage{data(initMeansC5D2a)}
\format{
  A matrix [nClusters x nFeatures] = [5 x 2].
}
\examples{
data(initMeansC5D2a); plot(initMeansC5D2a)
}
\keyword{datasets}
