% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RoughClustering.r
\name{normalizeMatrix}
\alias{normalizeMatrix}
\title{Matrix Normalization}
\usage{
normalizeMatrix( dataMatrix, normMethod, byrow)
}
\arguments{
\item{dataMatrix}{Matrix with the objects to be normalized.}

\item{normMethod}{1 = unity interval, 2 = normal distribution (sample variance), 3 = normal distribution (population variance). Any other value returns the matrix unchanged. Default: meansMatrix = 1 (unity interval).}

\item{byrow}{TRUE = rows are normalized, FALSE = columns are normalized. Default: byrow = TRUE (rows are normalized).}
}
\value{
Normalized matrix.
}
\description{
normalizeMatrix delivers a normalized matrix.
}
\author{
M. Goetz, G. Peters, Y. Richter, D. Sacker, T. Wochinger.
}

