% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoDataC2D2a.r
\docType{data}
\name{DemoDataC2D2a}
\alias{DemoDataC2D2a}
\title{A small two-dimensional dataset with two clusters for demonstration purposes. See examples in the Help/Description of a function, e.g. for HardKMeansDemo().}
\format{Rows: objects, columns: features}
\usage{
data(DemoDataC2D2a)
}
\description{
A small two-dimensional dataset with two clusters for demonstration purposes. See examples in the Help/Description of a function, e.g. for HardKMeansDemo().
}
\examples{
data(DemoDataC2D2a)
}
\keyword{datasets}
