% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMeansC2D2a.r
\docType{data}
\name{initMeansC2D2a}
\alias{initMeansC2D2a}
\title{Two-dimensional dataset with two initial cluster means for the dataset DemoDataC2D2a. See examples in the Help/Description of a function, e.g. for HardKMeansDemo().}
\format{Rows: objects, columns: features}
\usage{
data(initMeansC2D2a)
}
\description{
Two-dimensional dataset with two initial cluster means for the dataset DemoDataC2D2a. See examples in the Help/Description of a function, e.g. for HardKMeansDemo().
}
\examples{
data(initMeansC2D2a)
}
\keyword{datasets}
